requirements.txt
Django==1.4
PIL==1.1.7
PyYAML==3.09
Pygments==1.5
South==0.7.5
boto==2.0
bpython==0.11
django-annoying==0.7.6
django-envelope==0.4.1
django-extensions==0.8
django-honeypot==0.3.0
django-model-utils==1.1.0
django-storages==1.1.3
gunicorn==0.14.2
model-mommy==0.6.2
newrelic==1.2.1.265
python-dateutil==1.5
wsgiref==0.1.2
psycopg2==2.4.4
certifi==0.0.8
facepy==0.6.8
fandjango==4.0.6
requests==0.8.9
django-taggit==0.9.3
django-autoslug==1.5.0
django-floppyforms==0.4.7